[WIP] Support zlib streams over 2 GB
Final occurences, starting testing

+++ precomp.h
@@ -1,9 +1,9 @@
 int def(FILE *source, FILE *dest, int level, int windowbits, int memlevel);
-int def_compare(FILE *compfile, int level, int windowbits, int memlevel, int& decompressed_bytes_used, int decompressed_bytes_total, bool in_memory);
+long long def_compare(FILE *compfile, int level, int windowbits, int memlevel, long long & decompressed_bytes_used, long long decompressed_bytes_total, bool in_memory);
 long long def_compare_bzip2(FILE *source, FILE *compfile, int level, long long& decompressed_bytes_used);
 int def_part(FILE *source, FILE *dest, int level, int windowbits, int memlevel, long long stream_size_in, long long stream_size_out);
 int def_part_skip(FILE *source, FILE *dest, int level, int windowbits, int memlevel, long long stream_size_in, long long stream_size_out, int bmp_width);
-int inf(FILE *source, int windowbits, int& compressed_stream_size, int& decompressed_stream_size, bool& in_memory);
+int inf(FILE *source, int windowbits, long long& compressed_stream_size, long long& decompressed_stream_size, bool& in_memory);
 void zerr(int ret);
 #ifndef PRECOMPDLL
 #ifndef COMFORT
@@ -58,10 +58,10 @@ void show_used_levels();
 bool compress_file(float min_percent = 0, float max_percent = 100);
 void decompress_file();
 void convert_file();
-int try_to_decompress(FILE* file, int windowbits, int& compressed_stream_size, bool& in_memory);
-int try_to_decompress_bzip2(FILE* file, int compression_level, int& compressed_stream_size);
-void try_recompress(FILE* origfile, int comp_level, int mem_level, int windowbits, int& compressed_stream_size, long long decomp_bytes_total, bool in_memory);
-void try_recompress_bzip2(FILE* origfile, int level, int& compressed_stream_size);
+long long try_to_decompress(FILE* file, int windowbits, long long& compressed_stream_size, bool& in_memory);
+long long try_to_decompress_bzip2(FILE* file, int compression_level, long long& compressed_stream_size);
+void try_recompress(FILE* origfile, int comp_level, int mem_level, int windowbits, long long& compressed_stream_size, long long decomp_bytes_total, bool in_memory);
+void try_recompress_bzip2(FILE* origfile, int level, long long& compressed_stream_size);
 void write_header();
 void read_header();
 void convert_header();
@@ -130,7 +130,7 @@ public:
   }
 };
 
-void write_ftempout_if_not_present(int byte_count, bool in_memory, bool leave_open = false);
+void write_ftempout_if_not_present(long long byte_count, bool in_memory, bool leave_open = false);
 recursion_result recursion_compress(long long compressed_bytes, long long decompressed_bytes);
 recursion_result recursion_decompress(long long recursion_data_length);
